Add ReadOnly mode and make it default in the conf
ReadOnly mode is part of a database and is specified either in the
configuration or by command line flag. If either the flag --readonly or
the Database.ReadOnly option in the configuration is set to true, the
API will not allow database operations. This only applies to the API.

Additionally, the example configuration has been changed to bind on all
interfaces again, because ReadOnly is true by default.