Add verification by email
If SMTP is implicitly enabled (Conf.SMTP != nil), then an email is sent
via configured SMTP values to the email address provided when someone
registers a node. This contains a pseudorandom 64-bit integer (see #14),
which can be submitted to /api/verify in order to move the new node from
the nodes_verify_queue to the nodes table.

Close #8