Add node display and simplify index.html
The new node display uses the GeoJSON API to load the nodes. It accesses
`api/all?geojson` and uses the resulting object to populate the map. At
the moment, only nodes are loaded, without data such as IP and name.

Additionally, I have moved the script used for loading the map out of
index.html, and into its own file, loadmap.js, which could be
precompressed or minimized or something.

Ref #24