Add system for processing files at startup
On startup, the *fRes directory is now processed recursively and written
to a temporary directory. At the moment, this only inclues .tmpl files,
which are run through html/template with a configuration file as
data. This allows for single-time templating of static files.

In the future, this could include resource minimization and
precompression.

Additionally, there is a new configuration file field, Map, which
contains some map configuration options, such as center coordinates, zoom level, and tile server.

Close #25