Add periodic child map caching
This commit includes a helper function, UpdateMapCache(), which clears
the current cache and calls GetAllFromChildMaps() with the argument
being the new configuration option, ChildMaps.

This means that nodes from child maps will cache properly, and will be
refreshed on every heartbeat. It is done very inefficiently at the
moment, and map IDs are not garuenteed to be correct, but it is very
close to usable.

Ref #19