Added a white node.
Going to be inverting the color of the
node in the control when adding a new
node like when you click the distance
calculator. Also, I'll be getting rid
of the modal for adding a new node in
the near future if I can figure out what
I want to do with it.