Implement verify-time validity checking
This includes a system for adding additional validity checks, which can
be enabled or disabled in the configuration file. At the moment, the
default is to check that the IP a Node has is the one which verifies it.

In order to complete this, we will need to update the verification email
and include instructions for verifying from a remote server.

Ref #58