Removed contextmenu.
I feel that there's less of a need for a context menu now that we
have the navbar at the top, which explains everything that could
possibly be needing to be explained. If someone decides in the
future that they want to be able to have cerain right click
functions, then we can reopen this and go from there.

Closes #95