Send users to child maps if they want to message someone.
If there's a child map that has a node on the global map,
we can't have the users cross POST and cross GET data from
all of the domains that are being used safely, so it's best
to send them there and let them message from the child map
itself. Not only is it a less-safe, but it's browser
specific and not working in every browser, and support for
multiple browsers is more important than having two work.

Closes #127