Fix up Makefile so you don't need to 'make clean'
The whole problem was that the nodeatlas rule didn't depend on `.go`
files, so `make` would always consider it up to date, unless the
executable was missing. The solution was to add `$(wildcard *.go)` to
the rule declaration.

Thus, in order to recompile and run NodeAtlas, you can simply do

    make && ./nodeatlas

now. This will only recompile if a .go file has changed, and will
otherwise simply run the existing binary.