Update Makefile to remove stale assets
This was occasionally causing automatic redownload of every dependency
on invocation of `make`. Specifically, when a file was present in the
assets directory that was not updated by `getdeps.sh`, (such as an old
dependency that had since been removed from `depslist`), it would
maintain its old timestamp, causing `make` to try to update it again.

Now, all of the files matching the `DEPS` glob (`res/web/assets/*`,
meaning only visible files in that directory) are removed, and then
`getdeps.sh` is run.