Fixed issue with bootstrap compatability.
Originally, we didn't have to add form-control
to the form elements to get them styled, but
with the new version of bootstrap we did. The
past commit since we've upgraded bootstrap, I
had to just temporarily put some old bootstrap
code in the CSS file. That CSS is now gone.