Removed layer function code in preperation for new code.
Also I moved around a large function that should go at the bottom of
the file instead of the top because it was in the way of things. And
lastly, I added a "filter" variable at the top and an empty
`filterLayer` function that is going to be filled with code soon.