Fix bootstrap dependency problems
Let me take this moment to explain some things. The depslist largely
refers to minified JavaScript files hosted on GitHub. I won't defend the
practice of storing minified files in Git. It's dumb, and I wish there
was a better system.

However, when we refer to dependencies, we need to refer to a specific
version. That's why you see things like

    leaflet.markercluster.js https://raw.github.com/Leaflet/Leaflet.markercluster/0.2/dist/leaflet.markercluster.js

in the depslist. Note the `0.2`, which is a tag. That means it refers to
a specific Git commit ID. This is what's called a "ref", short for
"reference". Not all refs are tags, however - they can be commit IDs,
tags, branches, or even relatives, such as `HEAD^` or `master~4`.

Commit IDs and tags do not change. An ID is a SHA1 hash of a collection
of data about a commit, meaning that if you refer to such an ID, it is
guaranteed to always be the same commit. A tag is just an alias for a
commit ID.

Branches and relatives **do** change. That's their purpose; you commit
something on your branch, and git moves the branch ref forward.

That's why, if you're depending on retrieving a particular version of a
file via Git, you don't use branches. You don't use relatives. You use
either commit IDs or tags.

---

Before this commit, the bootstrap dependency was referring to the
`3.0.0-wip` **branch**. Recent commits on that branch caused some very
serious and hard to track down problems in NodeAtlas. People downloading
dependencies in the past several days would get a non-working version
with no clear explanation.