Fixed success message if NodeAtlas doesn't send mail.
There was already an option in the configuration file that asked if
NodeAtlas was to send mail or not, and we've been using that, but we
haven't done anything on the front end with that option until
now. Instead of the success message telling the user to check their
email if it doesn't send emails, it now just says that their node was
added. I also reformatted `res/web/js/node.js`.

Closes #160