Change PGP Disabled a little bit, and added a check.
It seems that there was a config option called PGP.PGPDisabled, but it
was set to `true` by default (which we don't want, incase someone
doesn't want PGP), and it was also *never* used. I added a check in
the Send func, and I also changed it from PGP.PGPDisabled to
PGP.Disabled.

Ref #165